Rename FixCacheKey -> CachedAs
